+++ rust/models/Cargo.toml
@@ -7,5 +7,13 @@ license = "GPL-2.0-or-later"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-serde = {version = "1", features = ["derive"]}
+serde = {version = "1", features = ["derive"], optional = true}
+#serde_json = "1"
+
+
+[features]
+default = ["serde_support"]
+serde_support = ["serde"]
+
+[dev-dependencies]
 serde_json = "1"
