+++ rust/models/src/credential.rs
@@ -1,38 +1,48 @@
-use serde::{Deserialize, Serialize};
-
 /// Represents a set of credentials to be used for scanning to access a host.
-#[derive(Serialize, Deserialize, Debug, Clone)]
+#[derive(Debug, Clone)]
+#[cfg_attr(
+    feature = "serde_support",
+    derive(serde::Serialize, serde::Deserialize)
+)]
 pub struct Credential {
     /// Service to use for accessing a host
     pub service: Service,
     /// Port used for getting access. If missing a standard port is used
     pub port: Option<u16>,
-    #[serde(flatten)]
+    #[cfg_attr(feature = "serde_support", serde(flatten))]
     /// Type of the credential to get access. Different services support different types.
     pub credential_type: CredentialType,
 }
 
 /// Enum of available services
-#[derive(Serialize, Deserialize, Debug, Clone)]
+#[derive(Debug, Clone)]
+#[cfg_attr(
+    feature = "serde_support",
+    derive(serde::Serialize, serde::Deserialize)
+)]
 pub enum Service {
-    #[serde(rename = "ssh")]
+    #[cfg_attr(feature = "serde_support", serde(rename = "ssh"))]
     /// SSH, supports [UP](CredentialType::UP) and [USK](CredentialType::USK) as credential types
     SSH,
-    #[serde(rename = "smb")]
+    #[cfg_attr(feature = "serde_support", serde(rename = "smb"))]
     /// SMB, supports [UP](CredentialType::UP)
     SMB,
-    #[serde(rename = "esxi")]
+    #[cfg_attr(feature = "serde_support", serde(rename = "esxi"))]
     /// ESXi, supports [UP](CredentialType::UP)
     ESXi,
-    #[serde(rename = "snmp")]
+    #[cfg_attr(feature = "serde_support", serde(rename = "snmp"))]
     /// SNMP, supports [SNMP](CredentialType::SNMP)
     SNMP,
 }
 
-#[derive(Debug, Deserialize, Serialize, Clone)]
+#[derive(Debug, Clone)]
+#[cfg_attr(
+    feature = "serde_support",
+    derive(serde::Serialize, serde::Deserialize)
+)]
 /// Enum representing the type of credentials.
 pub enum CredentialType {
-    #[serde(rename = "up")]
+    #[cfg_attr(feature = "serde_support", serde(rename = "up"))]
     /// User/password credentials.
     UP {
         /// The username for authentication.
@@ -40,18 +50,18 @@ pub enum CredentialType {
         /// The password for authentication.
         password: String,
     },
-    #[serde(rename = "usk")]
+    #[cfg_attr(feature = "serde_support", serde(rename = "usk"))]
     /// User/ssh-key credentials.
     USK {
         /// The username for authentication.
         username: String,
         /// The password for authentication.
         password: String,
-        #[serde(rename = "private")]
+        #[cfg_attr(feature = "serde_support", serde(rename = "private"))]
         /// The private key for authentication.
         private_key: String,
     },
-    #[serde(rename = "snmp")]
+    #[cfg_attr(feature = "serde_support", serde(rename = "snmp"))]
     /// SNMP credentials.
     SNMP {
         /// The SNMP username.
