+++ rust/models/src/parameter.rs
@@ -1,6 +1,8 @@
-use serde::{Deserialize, Serialize};
-
-#[derive(Serialize, Deserialize, Debug, Clone)]
+#[derive(Debug, Clone)]
+#[cfg_attr(
+    feature = "serde_support",
+    derive(serde::Serialize, serde::Deserialize)
+)]
 /// Represents a parameter for a VTS configuration.
 pub struct Parameter {
     /// The ID of the parameter.
