+++ rust/models/src/vt.rs
@@ -1,13 +1,18 @@
-use serde::{Deserialize, Serialize};
-
 use super::parameter::Parameter;
 
 /// A VT to execute during a scan, including its parameters
-#[derive(Serialize, Deserialize, Debug, Clone)]
+#[derive(Debug, Clone)]
+#[cfg_attr(
+    feature = "serde_support",
+    derive(serde::Serialize, serde::Deserialize)
+)]
 pub struct VT {
     /// The ID of the VT to execute
     pub oid: String,
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
+    #[cfg_attr(
+        feature = "serde_support",
+        serde(default, skip_serializing_if = "Vec::is_empty")
+    )]
     /// The list of parameters for the VT
     pub parameters: Vec<Parameter>,
 }
