+++ rust/scanner-api/src/routes.rs
@@ -3,8 +3,8 @@ use std::collections::HashMap;
 use crate::{
     error::APIError, guards::json_validation::JsonValidation, manager::ScanID, webserver::Manager,
 };
-use models::json::{
-    result::Result as ScanResult, scan::Scan, scan_action::ScanAction, status::Status as ScanStatus,
+use models::{
+    Result as ScanResult, Scan, ScanAction, Status as ScanStatus,
 };
 use rocket::{
     delete, get, head, http::Status, post, response::status::Created, serde::json::Json, uri, State,
